tech-lead: 修复测试基础设施和代码质量改进
- 修复 conftest.py: 添加 StaticPool 解决 in-memory SQLite 跨连接隔离问题
- 修复 config.py: 添加 JWT_SECRET_KEY 安全验证，防止不安全默认值
- 修复 models.py: datetime.utcnow → datetime.now(timezone.utc)
- 修复 credits.py: 原子化积分扣除（SQL WHERE 条件防竞态）
- 修复 chat.py: generator 内使用独立 SessionLocal 会话
- schemas.py: 添加 EmailStr 验证、密码/消息长度约束
- .env.example: 补充完整配置说明

+++ backend/app/schemas.py
@@ -23,7 +23,7 @@ class UserResponse(BaseModel):
     id: int
     email: str
     username: str
-    credits: float
+    credits: int
     created_at: datetime
 
     model_config = {"from_attributes": True}
@@ -31,13 +31,13 @@ class UserResponse(BaseModel):
 
 # Credits
 class CreditBalanceResponse(BaseModel):
-    credits: float
+    credits: int
 
 
 class CreditTransactionResponse(BaseModel):
     id: int
-    amount: float
-    balance_after: float
+    amount: int
+    balance_after: int
     type: str
     description: str
     created_at: datetime
